Fix windows build pt. 2

+++ src/cpp/src/sampling/sampler.cpp
@@ -952,10 +952,10 @@ SamplerOutput Sampler::sample(const std::vector<SequenceGroup::Ptr> & sequence_g
             if (!sampling_params.stop_strings.empty()) {
                 OPENVINO_ASSERT(m_tokenizer.m_pimpl != nullptr, "Stop strings require a valid tokenizer");
                 auto processed_stop_string = process_stop_strings(sampling_params.stop_strings, m_tokenizer);
-                m_stop_strings.insert({request_id, processed_stop_string});
+                m_stop_strings.insert({static_cast<int64_t>(request_id), processed_stop_string});
                 sequence_group->set_stream_window_size(processed_stop_string.first);
             } else {
-                m_stop_strings.insert({request_id, {size_t(0), {}}});
+                m_stop_strings.insert({static_cast<int64_t>(request_id), {size_t(0), {}}});
             }
         }
         const auto& stop_strings = m_stop_strings.at(request_id);
